qcacld-3.0: Optimize channel selection for NDI
In current algo maintains valid channel list for NDI from regulatory
channel list which leads to high space and time complexity.
As it is known that there are only three channels (144, 44, 6)
to support the NDI, so check for valid channel from these three
channels, instead of extracting from the channel list. If none of
above channels are valid, then take first valid channel from channel
list.

In this way, no need to maintain the channel list which helps to reduce
the time and space complexity.

Change-Id: I5dd2757ca952f4aaeb2a9ed74a47b4e547890245
CRs-Fixed: 3340778